requirements.txt
openai
pydub
requests
azure-functions
